Change Index and related typedefs from short to int to prevent overflow in simulations with over 2^16 extra bonds in the near future.

+++ src/structures.h
@@ -24,7 +24,7 @@ using namespace std;
 #define DrudeAtom        0x80
 
 
-typedef unsigned short Index;		//  Used for index into arrays
+typedef int Index;		//  Used for index into arrays
 					//  or parameters
 
 typedef struct atom_name_info
